mudança da avalição de precisão

+++ src/solSisLineares.hpp
@@ -15,7 +15,7 @@ const std::streamsize precisao_padrao = std::cout.precision();
 void quadrado(double& X, double& resultado);
 void derivada(void(*funcao)(double&, double&), double& ponto, double step, double& derivadaNoPonto);
 void newtonRaphson(void(*funcao)(double&, double&), double condicaoInicial,
-                   long long unsigned int maxIteracoes, double precisaoDaRaiz, double& resultado);
+                  uint maxIteracoes, double precisaoDaRaiz, double& resultado);
 void pontoFixo(void (*funcao)(double&, double&), double condicaoInicial,
                    uint maxIteracoes, double precisaoDaRaiz, double& resultado);
 void posicaoFalsa(void(*funcao)(double&, double&), double condicaoInicial_0, double condicaoInicial_1,
